updated QuoteTest.java to reflect set_up_GUI_for_feeds branch. --> all tests pass
commented out QuoteFeedByAnimeActivity.getAnimeTitles for now --> need to work on that!
